requirements.txt
Biopython==1.83
pandas
matplotlib
seaborn
scikit-learn
scipy
openpyxl